requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.18
openai>=1.6.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.2
python-dotenv==1.0.0
pillow==10.1.0

